Move `bootstrap/bootstrap_server` to `transport/tcp_server`
